Quick hotfixes (#75)
* Quick bug fixes

Forgot to remove references to serializer after a changed import
statement. Test runner will now not error if no media files were
created.

+++ cognoma_site/test_runner.py
@@ -23,7 +23,8 @@ class TempMediaMixin(object):
         Delete temp storage.
         """
         super(TempMediaMixin, self).teardown_test_environment()
-        shutil.rmtree(settings.MEDIA_ROOT)
+        if os.path.isdir(settings.MEDIA_ROOT):
+            shutil.rmtree(settings.MEDIA_ROOT)
 
 
 class TemporaryMediaTestSuiteRunner(TempMediaMixin, DiscoverRunner):
